+++ src/main/kotlin/com/nanovms/ops/Service.kt
@@ -3,6 +3,7 @@ package com.nanovms.ops
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
+import com.nanovms.ops.command.Command
 
 interface Service: Disposable {
     fun runExecutable(filepath: String, configpath: String? = null): OpsResult<String>
