+++ src/main/kotlin/com/nanovms/ops/command/BuildCommand.kt
@@ -1,7 +1,6 @@
-package com.nanovms.ops.action
+package com.nanovms.ops.command
 
 import com.intellij.openapi.project.Project
-import com.nanovms.ops.Command
 
 class BuildCommand(project: Project, private val filepath: String): Command(project) {
     override fun createArguments(): Collection<String> {
