+++ src/main/kotlin/com/nanovms/ops/command/Command.kt
@@ -1,8 +1,9 @@
-package com.nanovms.ops
+package com.nanovms.ops.command
 
 import com.intellij.execution.process.OSProcessHandler
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
+import com.nanovms.ops.Service
 import java.io.File
 
 abstract class Command(val project: Project) {
